Fix five defects found dogfooding an invoicing app

+++ examples/shop/migrations/20260822171947-create-products.up.sql
@@ -1,10 +1,10 @@
--- Migration 20260824045114: Create products table
+-- Migration 20260919100909: Create products table
 
 CREATE TABLE IF NOT EXISTS products (
   id UUID PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   sku VARCHAR(255) NOT NULL,
-  price DOUBLE PRECISION,
+  price DECIMAL(19,4),
   created_at TIMESTAMP NOT NULL,
   updated_at TIMESTAMP,
   deleted_at TIMESTAMP
